refacto(store): adding an interface C4ElementStateHandler to match the different event handlers of the C4 layers and implementing it with the System element
